useTaskList was added into all needed components
